hasy-symbol-api/requirements.txt.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
